% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{flip}
\alias{flip}
\title{Flip Rows of a Matrix or Other Object}
\usage{
flip(x, ...)
}
\arguments{
\item{x}{object to be flipped}
}
\value{
reverses most objects, e.g. lists, reverses the rows of a matrix
}
\description{
Flip Rows of a Matrix or Other Object
}
