% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gell-package.R
\docType{package}
\name{gell-package}
\alias{gell}
\alias{gell-package}
\title{gell: Draw and Manipulate Ellipses}
\description{
Draw and manipulate ellipses using projective geometry.
}
\author{
\strong{Maintainer}: Georges Monette \email{georges@yorku.ca}

}
\keyword{internal}
