% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\docType{package}
\name{gell}
\alias{gell}
\alias{gell-package}
\title{gell: Working with Ellipses}
\usage{
gell(x, ...)
}
\arguments{
\item{x}{an object of class ell to be transformed to a gell, or}

\item{center}{the center of the gell, default c(0,0), plus some combination of the following:}

\item{shape}{a \eqn{2 \times 2} matrix giving the variance corresponding to the ellipse}

\item{A}{a \eqn{2 \times 2} matrix with the transformation of the unit circle generating the shape
of the ellipse}

\item{form}{a \eqn{2 \times 2} matrix with the matrix of the quadratic form, i.e. the inverse of \code{shape}}

\item{gamma}{an orthonormal matrix corresponding to the eigenvectors of 'shape'}

\item{d}{the eigenvalues of 'shape'}

\item{radius}{a factor by which to scale the ellipse as defined by other parameters}
}
\description{
This is a part of a package for plotting objects
represented in projective 2 dimensional space
Using objects 'point' and 'line'

\subsection{Generalized ellipses}{
}
}
\details{
General approach

See summary and examples at end

Plotting Functions:
\itemize{
\item \code{\link[=pl]{pl()}}  a method that plots most objects
\item \code{\link[=text]{text()}} a method that adds text at a location
\item \link{text} at a point
\item \link{text(point)} at a point
\item \link{contour.function} S3 method using \link[graphics:contour]{graphics::contour}
}

Methods:
\itemize{
\item fta: 'from', 'to', 'along'
\item axes: gell, point (dir), type
}

Idea:
\itemize{
\item Use the projective representation of points and lines
to include the points at infinity and the line at infinity
thereby making it easy to work with parallel lines and, e.g.
the line joining a point with the intersection of two lines
that happen to be parallel, etc.
\item Manipulated projective points and lines with a suite of
operators that coerce finite points and lines to projective
form.
\item We need three classes of objects (but class might be implicit)
\itemize{
\item ppoints
\item plines (unbounded)
\item psegs  for segments and, e.g. parallelograms, etc.
}
\item A sequence of n points can be intererpreted as n-1 segments
or as n-1 lines
\item A sequence of n lines can generate n-1 intersections
\item Generalized ellipses, class 'gell'
For now represented using a svd representation using
orthogonal \eqn{$\Gamma$} and singular values which may be
0 or Inf.
\item Write plotting methods for base R and for lattice that
plot projective points and lines and gell's. Note that lines can be
drawn with 'abline' to avoid problems with endpoints.
\item Use projective geometry to easily compute the intersection of two lines
and the line joining two points
\item Issues:
\itemize{
\item Lines can be represented as a projective line (3-vector) or as
two endpoints.  The former needs to be transformed to the latter
for actual plotting.
\item We need to distinguish between a line segment and a line.
\itemize{
\item lines could be plotted with \code{abline} so no endpoints
\item line segments can be turned into lines
}
}
}

Operators:
\itemize{
\item 1 or more lines
\itemize{
\item intersection
\item copuntal
\item rotate
\item transform
\item 
\item dual point of a line (o a finite line = point at inf perpendicular)
}
\item line and point
\itemize{
\item perpendicular line through point (might add ell to get conjugate line)
\item Relation: point in line?
}
\item 2 or more points
\itemize{
\item sum
\item scalar
}
\item \code{\link[=gell]{gell()}} and
\itemize{
\item point
\itemize{
\item scale to gell
\item conjugate \code{\link[=axes]{axes()}}
\item conjugate parallelogram
\item subtending rectangle
}
}
\item gell and
\itemize{
\item line
\itemize{
\item tangent point
}
}
\item gell and gell
\itemize{
\item locus of osculation
}
}

Constructor for gell objects
}
\author{
\strong{Maintainer}: Georges Monette \email{georges@yorku.ca}

}
\keyword{internal}
