% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{init}
\alias{init}
\title{Initialize plotting surface}
\usage{
init(
  xlim = fac * c(-1, 1),
  ylim = xlim,
  fac = 3,
  xlab = "",
  ylab = "",
  asp = 1,
  ...
)
}
\arguments{
\item{fac}{expand c(-1,1)}

\item{xlab}{x label}

\item{ylab}{y label}

\item{asp}{(default: 1)}

\item{from}{lower left corner}

\item{to}{upper right corned}
}
\description{
Initialize plotting surface
}
