% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{inter}
\alias{inter}
\alias{inter,gell,gell,missing-method}
\alias{inter,gell,gell,numeric-method}
\alias{inter,gell,line,missing-method}
\alias{inter,gell,seg,missing-method}
\alias{inter,gell,point,missing-method}
\alias{inter,function,numeric,missing-method}
\alias{inter,line,line,missing-method}
\alias{inter,point,point,missing-method}
\title{Generate the intersection of two objects}
\usage{
inter(x, y, option, ...)

\S4method{inter}{gell,gell,missing}(x, y, option, ...)

\S4method{inter}{gell,gell,numeric}(x, y, option, ...)

\S4method{inter}{gell,line,missing}(x, y, option, ...)

\S4method{inter}{gell,seg,missing}(x, y, option, ...)

\S4method{inter}{gell,point,missing}(x, y, option, ...)

\S4method{inter}{`function`,numeric,missing}(x, y, option, ...)

\S4method{inter}{line,line,missing}(x, y, option, ...)

\S4method{inter}{point,point,missing}(x, y, option, ...)
}
\arguments{
\item{x}{gell or function}

\item{y}{gell, line, segment, point or numeric}

\item{option}{fineness of grid over which to find intersections, default:}
}
\value{
intersecting point(s)
}
\description{
Generate the intersection of two objects
}
\section{Methods (by class)}{
\itemize{
\item \code{inter(x = gell, y = gell, option = missing)}: intersection of two \code{gell} objects using \link{inter_gells}

\item \code{inter(x = gell, y = gell, option = numeric)}: intersection of two \code{gell} objects

\item \code{inter(x = gell, y = line, option = missing)}: intersection a \code{gell} object with a \code{line}

\item \code{inter(x = gell, y = seg, option = missing)}: intersection a \code{gell} object with a \code{seg} object (segment)

\item \code{inter(x = gell, y = point, option = missing)}: intersection a \code{gell} object with a \code{point}

\item \code{inter(x = `function`, y = numeric, option = missing)}: intersection a \code{function} object with a numeric value(s)

\item \code{inter(x = line, y = line, option = missing)}: intersection of two \code{line} objects

\item \code{inter(x = point, y = point, option = missing)}: intersection of two \code{point} objects, i.e. the 'join' of two points

}}
