% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{inter_ell_line}
\alias{inter_ell_line}
\title{Find intersection of ellipse with line}
\usage{
inter_ell_line(ell, line)
}
\arguments{
\item{ell}{an object of class ell}

\item{line}{an object of class line}
}
\description{
Find intersection of ellipse with line
}
