% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{inter_gell_line}
\alias{inter_gell_line}
\title{Find intersection of a generalized ellipse with line}
\usage{
inter_gell_line(ge, line)
}
\arguments{
\item{ge}{an object of class gell}

\item{line}{an object of class line}
}
\description{
Find intersection of a generalized ellipse with line
}
