% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{inter_gells}
\alias{inter_gells}
\title{Intersecting ellipses}
\usage{
inter_gells(g1, g2, n = 1000, verbose = 0)
}
\arguments{
\item{g1}{gell object}

\item{g2}{gell object}

\item{n}{fineness of grid to compute intersections}

\item{verbose}{default: 0}
}
\description{
Normally called from S4 method \link{inter}
}
