% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{join}
\alias{join}
\title{S4 Generic Function for the Join of Projective Lines and Points}
\usage{
join(x, y, ...)
}
\arguments{
\item{x, y}{two projective lines or two projective points}
}
\value{
the line joining two points or the intersection of
two lines. The the lines are parallel the intersection
is a point at infinity.
}
\description{
The intersection of two lines or the line generated by two points
}
\examples{
join(p(c(1,1),c(1,-1)), p(c(2,2)))
join(
  p(1,1,0),  # point at infinity)
  p(c(1,2))
)
join(p(c(1,1,1)), p(c(1,1,1))) \%>\% join(l(c(1,-1,0))) \%>\% join(p(c(1,1,1)))
}
