% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{l}
\alias{l}
\title{Constructor for line object}
\usage{
l(x, ...)
}
\arguments{
\item{x, ...}{one of the following:
\itemize{
\item a vector of length 2 representing a line through the origin. The 45 degree
line is represented by (1, -1, 0), i.e. the points (x, y) satisfying 1*x + (-1)*y + 0 = 0.
\item a vector of length 3: if the third element is not zero: a line not through
the origin point in homegeneous coordinates or,
if the third element is 0: a point at infinity, i.e. a direction.
\item a n by 2 matrix of affine points.
\item a n by 3 matrix of both affine points and points at infinity
\item objects that can be interpreted as projective points, e.g. objects of class 'seg'
}}
}
\value{
An object of class \code{line} with the structure of an n by 3 matrix where each row
represents a line.
}
\description{
Lines are represented by rows of n x 3 matrices where
each row is the projective representation of a 2D line
in homogeneous coordinates, or the representation of the
line at infinity, (0, 0, c) where c is not zero.
}
\details{
The row vector (a, b, c) represents the line consisting
of pairs (x, y) such that ax + by + c = 0.

Note that if c = 0, the line goes through the origin.

The line at infinity is represented by the vector (0, 0, c),
with c != 0.

The vector (0, 0, 0) does not define a line.
}
