% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{n}
\alias{n}
\title{Normalize homogeneous coordinates or gell}
\usage{
n(x, ...)
}
\arguments{
\item{x}{an object represented as a n x 3 numerical matrix}
}
\description{
Normalize homogeneous coordinates or gell
}
