% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{p}
\alias{p}
\title{Constructor for point object}
\usage{
p(x, ...)
}
\arguments{
\item{x, ...}{one of the following:
\itemize{
\item a vector of length 2 representing an affine point (i.e. not a point at infinity)
\item a vector of length 3: if the third element is not zero: an affine point in homegeneous coordinates or,
if the third element is 0: a point at infinity, i.e. a direction.
\item a n by 2 matrix of affine points.
\item a n by 3 matrix of both affine points and points at infinity
\item objects that can be interpreted as projective points, e.g. objects of class 'seg'
}}
}
\value{
An object of class \code{line} with the structure of an n by 3 matrix where each row
represents a line.
}
\description{
Points are represented by rows of n x 3 matrices where
each row is the projective representation of a point
in homogeneous coordinates.
}
\details{
The row vector (a, b, c) represents the point consisting of
the affine point (a/c, b/c) if c is not equal to 0, or
the point at infinity (a, b, 0) representing the 'direction'
pointing from the origin through the point (a, b).

Two parallel lines meet at a point at infinity

The vector (0, 0, 0) does not define a point.
}
