% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{perp}
\alias{perp}
\title{Drop a perpendicular}
\usage{
perp(x, ...)
}
\arguments{
\item{x, ...}{line and point or point and line to drop perpendicular
from point to line}
}
\description{
Drop a perpendicular
}
\examples{
init()
join(p(-1,-1),p(-2,2)) \%>\% pl -> line1
p(p(-1,-1),p(-2,2)) \%>\% pl
perp(line1) \%>\% pl(col = 'red')
p(0,0) \%>\% pl(pch=16)
perp(line1,p(2,2)) \%>\% pl(col = 'red')

}
\seealso{
\code{\link[=fta]{fta()}}
}
