% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{pl}
\alias{pl}
\alias{pl.ell}
\alias{pl.gell}
\title{Plot generic function}
\usage{
pl(x, ...)

pl(x, ...)

\method{pl}{ell}(ell, ...)

\method{pl}{gell}(gell, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{plotting parameters}
}
\description{
Plot generic function

S3 Plotting Generic Function for Graphic Objects in the Gell Package
}
\section{Methods (by class)}{
\itemize{
\item \code{pl(ell)}: method for objects of class 'ell'

\item \code{pl(gell)}: method for objects of class 'ell'

}}
