% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{pl.seg}
\alias{pl.seg}
\title{Plot segments}
\usage{
\method{pl}{seg}(x, close = FALSE, type = "l", fill = NULL, verbose = 0, ...)
}
\arguments{
\item{close}{if TRUE repeat first point at end of sequence (default: FALSE)}

\item{type}{'p', 'b' or 'l' (default)}

\item{fill}{fill color (default: NULL)}

\item{...}{other parameters passed to plotting function}
}
\description{
Plot segments
}
