% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{rot}
\alias{rot}
\title{ell ####
probably no longer needed}
\usage{
rot(m, theta, ...)
}
\arguments{
\item{m}{object to be rotated}

\item{theta}{angle by which to rotate object counterclockwise}
}
\description{
Rotate an object
}
