% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{close.seg}
\alias{close.seg}
\title{Close a sequence of segments}
\usage{
\method{close}{seg}(sg)
}
\arguments{
\item{sg}{segments to be terminated with first point}
}
\description{
Terminate a sequence of segments with the first point in the sequence
}
\section{Functions}{
\itemize{
\item \code{close(seg)}: close sequence of points by ending with first point

}}
