% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{contour.function}
\alias{contour.function}
\title{Plot contour lines for a function}
\usage{
\method{contour}{`function`}(
  fun,
  val = 0,
  xlim = expand(par("usr")[1:2], fac),
  ylim = expand(par("usr")[3:4], fac),
  fac = 1.1,
  n = 200
)
}
\description{
Uses \link[graphics:contour]{graphics::contour} to draw contour lines
}
