% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projective.R
\name{plus}
\alias{plus}
\alias{mult}
\title{S4 methods to add a point or multiply objects by a scalar}
\usage{
plus(x, y, ...)

mult(x, y, ...)
}
\arguments{
\item{x, y}{lines, points, gells or matrices to be added to
each other or to be displaced by a point}
}
\description{
S4 methods to add a point or multiply objects by a scalar
}
\section{Functions}{
\itemize{
\item \code{mult()}: multiply object by a scalar

}}
